## 2018-06-10 System Themes * Start work on getting the system themes to display the extras. Issue #37 & #38   * Not perfect but at least the "Chrome" displays the intended buttons and text(s). * Removal of "not needed" casts.

+++ Source/Krypton Components/ComponentFactory.Krypton.Docking/Control Docking/KryptonAutoHiddenSlidePanel.cs
@@ -556,7 +556,7 @@ namespace ComponentFactory.Krypton.Docking
                             {
                                 cbSize = (uint)Marshal.SizeOf(typeof(PI.TRACKMOUSEEVENTS)),
                                 dwHoverTime = 100,
-                                dwFlags = (int)(PI.TME_LEAVE),
+                                dwFlags = PI.TME_LEAVE,
                                 hWnd = Handle
                             };
